requirements.txt
surfboard
docutils
matplotlib
torchvision
praat-parselmouth
tqdm
librosa
Sphinx
seaborn
Unidecode
tensorflow_gpu
sounddevice
scipy
torch
inflect
numpy
cxxfilt
scikit_learn
sphinx_rtd_theme
tensorboardX
tensorflow
